+++ src/main/java/com/board_of_ads/service/impl/AuthVkServiceImpl.java
@@ -1,25 +1,42 @@
 package com.board_of_ads.service.impl;
 
+import com.board_of_ads.models.Image;
+import com.board_of_ads.models.User;
 import com.board_of_ads.service.interfaces.AuthVkService;
+import com.board_of_ads.service.interfaces.UserService;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.JSONValue;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
+import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
 
 @Service
+@Getter
 @Setter
+@RequiredArgsConstructor
 @ConfigurationProperties(prefix="security.auth-vk")
 public class AuthVkServiceImpl implements AuthVkService {
+
+    private final UserService userService;
+
 
     private String clientId;
     private String clientSecret;
@@ -33,6 +50,14 @@ public class AuthVkServiceImpl implements AuthVkService {
     private String fields;
     private String version;
 
+
+    public void auth(String code) {
+        String response = getAuthResponseURL(code);
+        Map<String, String> userData = getUserData(response);
+        userData = getUserData(userData);
+        User user = init(userData);
+        login(user);
+    }
     /**
      * Метод для кнопки авторизации
      * @return ссылку для авторизации при переходе по которой, получим ссылку с
@@ -54,8 +79,6 @@ public class AuthVkServiceImpl implements AuthVkService {
      */
     @Override
     public String getAuthResponseURL(String code) {
-        System.out.println(clientId);
-        System.out.println(clientSecret);
         return tokenURL + "?"
                 + "client_id=" + clientId
                 + "&client_secret=" + clientSecret
@@ -112,4 +135,39 @@ public class AuthVkServiceImpl implements AuthVkService {
         userData.put("avatar_link", (String) dataArray.get("photo_100"));
         return userData;
     }
+
+    /**
+     * Метод для получения сессии пользователя
+     */
+    @Override
+    public void login(User user) {
+        Authentication authentication = new UsernamePasswordAuthenticationToken(user, null, user.getAuthorities());
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+
+    /**
+     * Метод инициализации пользователя.
+     * Если такой пользователь есть в базе данных, то он вернет его.
+     * Если пользователя не существует, то он его создаст, добавит в БД и вернет.
+     *
+     * @param userData возвращается методом getUserData
+     */
+    @Override
+    public User init(Map<String, String> userData) {
+        User user = userService.getUserByEmail(userData.get("email"));
+        if (user != null) {
+            return user;
+        }
+        user = new User();
+        user.setEnable(true);
+        user.setDataRegistration(LocalDateTime.now());
+        user.setEmail(userData.get("email"));
+        user.setFirsName(userData.get("first_name"));
+        user.setLastName(userData.get("last_name"));
+        user.setAvatar(new Image(null, userData.get("avatar_link")));
+        user.setPassword(userData.get("email")); //todo create set password page (and phone)
+        userService.saveUser(user);
+        return user;
+    }
 }
+
