+++ src/main/java/com/board_of_ads/service/interfaces/AuthVkService.java
@@ -1,8 +1,12 @@
 package com.board_of_ads.service.interfaces;
 
+import com.board_of_ads.models.User;
+
 import java.util.Map;
 
 public interface AuthVkService {
+
+    void auth(String code);
 
     String getAuthURL();
 
@@ -11,4 +15,8 @@ public interface AuthVkService {
     Map<String, String> getUserData(String authResponseUrl);
 
     Map<String, String> getUserData(Map<String, String> userData);
+
+    void login(User user);
+
+    User init(Map<String, String> userData);
 }
